Add view elements needed for chat
