+++ app/src/main/java/com/silverpants/instantaneous/data/chat/model/Message.kt
@@ -1,0 +1,9 @@
+package com.silverpants.instantaneous.data.chat.model
+
+import java.util.*
+
+data class Message(
+    val userId: String,
+    val message: String,
+    val timestamp: Date,
+)
