+++ app/src/main/res/layout/item_chat_new.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_margin="@dimen/size_m"
@@ -12,23 +11,24 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/grey_300"
+        android:background="@color/blue"
         android:minHeight="@dimen/chat_height_min">
 
-        <TextView
-            android:id="@+id/tv_chat_me_text"
+        <EditText
+            android:id="@+id/et_chat_new"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_margin="@dimen/size_m"
-            android:textColor="@color/black"
+            android:background="@null"
+            android:hint="Type something here...."
+            android:textColor="@color/white"
             android:textSize="16sp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.2"
+            app:layout_constraintHorizontal_bias="0.6"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintWidth_max="0dp"
-            tools:text="Nothing much, Chillin" />
+            app:layout_constraintWidth_max="0dp" />
     </androidx.constraintlayout.widget.ConstraintLayout>
